Refactor routes into modules
