+++ app/dependencies.py
@@ -1,0 +1,8 @@
+from . import database
+
+def get_db():
+    db = database.SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
